refactor(config): update a few more config imports
